requirements.txt
aniso8601==8.0.0
appdirs==1.4.4
atomicwrites==1.4.0
attrs==20.2.0
autocorrect==2.2.2
azure-ai-textanalytics==5.1.0b2
azure-common==1.1.25
azure-core==1.8.2
black==20.8b1
breadability==0.1.20
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
clean-text==0.3.0
click==7.1.2
clickclick==20.10.2
colorama==0.4.4
connexion==2.7.0
cryptography==3.2.1
docopt==0.6.2
emoji==0.6.0
fastapi==0.62.0
Flask==1.1.2
Flask-Cors==3.0.9
flask-restplus==0.13.0
ftfy==5.8
h11==0.11.0
idna==2.10
inflection==0.5.1
iniconfig==1.1.1
install==1.3.4
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
lxml==4.6.1
MarkupSafe==1.1.1
msrest==0.6.19
mypy-extensions==0.4.3
nltk==3.5
numpy==1.19.3
oauthlib==3.1.0
openapi-spec-validator==0.2.9
packaging==20.7
pathspec==0.8.1
pluggy==0.13.1
py==1.9.0
pycountry==20.7.3
pycparser==2.20
pydantic==1.7.3
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent==0.17.3
pytest==6.1.2
python-docx==0.8.10
python-dotenv==0.15.0
python-multipart==0.0.5
pytz==2020.4
PyYAML==5.3.1
regex==2020.10.28
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sortedcontainers==2.2.2
soupsieve==2.0.1
starlette==0.13.6
sumy==0.8.1
swagger-ui==0.1.2
swagger-ui-bundle==0.0.8
tika==1.24
toml==0.10.2
tqdm==4.51.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.11
uvicorn==0.12.3
wcwidth==0.2.5
Werkzeug==1.0.1
wikipedia==1.4.0
